+++ frontend/src/admin/tasks/index.ts
@@ -1,87 +1,17 @@
 import { hyperStyled } from "@macrostrat/hyper";
-import { useRef, useMemo, useEffect, useState } from "react";
+import { useRef, useMemo, useState } from "react";
 import { CollapseCard } from "@macrostrat/ui-components";
 import { useAPIv2Result } from "~/api-v2";
-import useResizeObserver from "use-resize-observer";
 import { MinimalNavbar, ServerStatus } from "~/components";
 import { Button, Menu, MenuItem, NonIdealState } from "@blueprintjs/core";
 import styles from "./module.styl";
-import { FixedSizeList as List } from "react-window";
 const h = hyperStyled(styles);
 import { useParams, Switch, Route } from "react-router";
-import { Link } from "react-router-dom";
-import { parse } from "ansicolor";
-import classNames from "classnames";
+import { Link, useHistory } from "react-router-dom";
 import Form from "@rjsf/core";
 import { useSparrowWebSocket } from "~/api-v2";
-
-const Row = ({ data, index, style }) => {
-  const lineno = index - 1;
-  const txt = data[lineno] ?? "";
-  const spans = parse(txt).spans;
-
-  return h(
-    "div.message",
-    {
-      style: style,
-    },
-    [
-      h.if(lineno >= 0 && lineno < data.length)(
-        "span.lineno.dark-gray",
-        `${lineno + 1}`
-      ),
-      h(
-        "span.message-text",
-        spans.map((d) => {
-          const { italic, bold, text, color, bgColor } = d;
-          const bg = bgColor?.name != null ? `bg-${bgColor.name}` : null;
-          const className = classNames(
-            {
-              italic,
-              bold,
-              dim: color?.dim,
-              "bg-dim": bgColor?.dim,
-            },
-            color?.name,
-            bg
-          );
-          return h("span", { className }, text);
-        })
-      ),
-    ]
-  );
-};
-
-const LogWindow = ({ messages }) => {
-  const ref = useRef<List>();
-  const {
-    ref: containerRef,
-    width = 1,
-    height = 1,
-  } = useResizeObserver<HTMLDivElement>();
-  const extraLines = 2;
-  useEffect(() => {
-    ref.current?.scrollToItem(messages.length + extraLines);
-  }, [messages.length]);
-
-  return h(
-    "div.log-window",
-    { ref: containerRef },
-    h(
-      List,
-      {
-        ref,
-        height,
-        itemCount: messages.length + extraLines,
-        itemSize: 18,
-        itemData: messages,
-        width,
-        className: "message-history",
-      },
-      Row
-    )
-  );
-};
+import { LogWindow } from "./log-window";
+import { LinkButton } from "@macrostrat/ui-components/lib/esm/ext/router-links";
 
 function getDefaultsForSchema(schema) {
   let defaults = {};
@@ -213,17 +143,34 @@ const TasksRouter = ({ base, tasks }) => {
   ]);
 };
 
-function TaskList({ tasks, base, task }) {
+function TaskMenuItem({ taskName, base, isActive }) {
+  const history = useHistory();
+
+  const linkDest = base + `/${taskName}`;
+  return h(MenuItem, {
+    active: isActive,
+    selected: isActive,
+    text: h(
+      Link,
+      {
+        to: linkDest,
+      },
+      taskName
+    ),
+    onClick(e) {
+      history.push(linkDest);
+      e.preventDefault();
+    },
+  });
+}
+
+function TaskList({ tasks = [], base, task }) {
   return h("div.tasks-list", [
     h("h3", "Tasks"),
     h(
       Menu,
-      (tasks ?? []).map((d) =>
-        h(MenuItem, {
-          text: h(Link, { to: base + `/${d.name}` }, d.name),
-          selected: d.name === task,
-          disabled: d.name === task,
-        })
+      tasks.map((d) =>
+        h(TaskMenuItem, { taskName: d.name, base, isActive: d.name === task })
       )
     ),
   ]);
