+++ frontend/src/admin/tasks/module.styl
@@ -36,8 +36,10 @@
   font-size 14px
 
 .left-column
-  max-width 20em
+  max-width 24em
   margin-right 1em
+  .tasks-list :global(.bp3-button)
+    text-align: left
 
 .log-window
   flex-grow 1
